--- a/src/modi/c13.rs
+++ b/src/modi/c13.rs
@@ -1,4 +1,4 @@
-use scroll::{ctx::TryFromCtx, Pread};
+use scroll::{ctx::TryFromCtx, Endian, Pread};
 
 use crate::common::*;
 use crate::modi::{constants, FileChecksum, FileIndex, FileInfo, LineInfo, LineInfoKind};
@@ -40,12 +40,26 @@ impl DebugSubsectionKind {
     }
 }
 
-#[derive(Clone, Copy, Debug, Pread)]
+#[derive(Clone, Copy, Debug)]
 struct DebugSubsectionHeader {
     /// The kind of this subsection.
     kind: u32,
     /// The length of this subsection in bytes, following the header.
     len: u32,
+}
+
+impl<'t> TryFromCtx<'t, Endian> for DebugSubsectionHeader {
+    type Error = scroll::Error;
+    type Size = usize;
+
+    fn try_from_ctx(this: &'t [u8], le: Endian) -> scroll::Result<(Self, Self::Size)> {
+        let mut offset = 0;
+        let data = DebugSubsectionHeader {
+            kind: this.gread_with(&mut offset, le)?,
+            len: this.gread_with(&mut offset, le)?,
+        };
+        Ok((data, offset))
+    }
 }
 
 impl DebugSubsectionHeader {
@@ -97,10 +111,23 @@ impl<'a> FallibleIterator for DebugSubsectionIterator<'a> {
     }
 }
 
-#[derive(Clone, Copy, Debug, Default, Pread)]
+#[derive(Clone, Copy, Debug, Default)]
 struct DebugInlineeLinesHeader {
     /// The signature of the inlinees
     signature: u32,
+}
+
+impl<'t> TryFromCtx<'t, Endian> for DebugInlineeLinesHeader {
+    type Error = scroll::Error;
+    type Size = usize;
+
+    fn try_from_ctx(this: &'t [u8], le: Endian) -> scroll::Result<(Self, Self::Size)> {
+        let mut offset = 0;
+        let data = DebugInlineeLinesHeader {
+            signature: this.gread_with(&mut offset, le)?,
+        };
+        Ok((data, offset))
+    }
 }
 
 impl DebugInlineeLinesHeader {
@@ -194,7 +221,7 @@ impl<'a> DebugInlineeLinesSubsection<'a> {
     }
 }
 
-#[derive(Clone, Copy, Debug, Default, Pread)]
+#[derive(Clone, Copy, Debug, Default)]
 struct DebugLinesHeader {
     /// Section offset of this line contribution.
     offset: PdbInternalSectionOffset,
@@ -202,6 +229,21 @@ struct DebugLinesHeader {
     flags: u16,
     /// Code size of this line contribution.
     code_size: u32,
+}
+
+impl<'t> TryFromCtx<'t, Endian> for DebugLinesHeader {
+    type Error = scroll::Error;
+    type Size = usize;
+
+    fn try_from_ctx(this: &'t [u8], le: Endian) -> scroll::Result<(Self, Self::Size)> {
+        let mut offset = 0;
+        let data = DebugLinesHeader {
+            offset: this.gread_with(&mut offset, le)?,
+            flags: this.gread_with(&mut offset, le)?,
+            code_size: this.gread_with(&mut offset, le)?,
+        };
+        Ok((data, offset))
+    }
 }
 
 impl DebugLinesHeader {
@@ -242,7 +284,7 @@ enum LineMarkerKind {
 
 /// The raw line number entry in a PDB.
 #[repr(C, packed)]
-#[derive(Clone, Copy, Debug, Pread)]
+#[derive(Clone, Copy, Debug)]
 struct LineNumberHeader {
     /// Offset to start of code bytes for line number.
     offset: u32,
@@ -254,6 +296,20 @@ struct LineNumberHeader {
     /// unsigned long   fStatement  :1;   // true if a statement line number, else an expression
     /// ```
     flags: u32,
+}
+
+impl<'t> TryFromCtx<'t, Endian> for LineNumberHeader {
+    type Error = scroll::Error;
+    type Size = usize;
+
+    fn try_from_ctx(this: &'t [u8], le: Endian) -> scroll::Result<(Self, Self::Size)> {
+        let mut offset = 0;
+        let data = LineNumberHeader {
+            offset: this.gread_with(&mut offset, le)?,
+            flags: this.gread_with(&mut offset, le)?,
+        };
+        Ok((data, offset))
+    }
 }
 
 /// A mapping of code section offsets to source line numbers.
@@ -358,11 +414,25 @@ impl FallibleIterator for DebugLinesIterator<'_> {
     }
 }
 
-#[derive(Clone, Copy, Debug, Default, Pread)]
+#[derive(Clone, Copy, Debug, Default)]
 #[repr(C, packed)]
 struct ColumnNumberEntry {
     start_column: u16,
     end_column: u16,
+}
+
+impl<'t> TryFromCtx<'t, Endian> for ColumnNumberEntry {
+    type Error = scroll::Error;
+    type Size = usize;
+
+    fn try_from_ctx(this: &'t [u8], le: Endian) -> scroll::Result<(Self, Self::Size)> {
+        let mut offset = 0;
+        let data = ColumnNumberEntry {
+            start_column: this.gread_with(&mut offset, le)?,
+            end_column: this.gread_with(&mut offset, le)?,
+        };
+        Ok((data, offset))
+    }
 }
 
 #[derive(Clone, Debug, Default)]
@@ -385,7 +455,7 @@ impl FallibleIterator for DebugColumnsIterator<'_> {
 }
 
 #[repr(C, packed)]
-#[derive(Clone, Copy, Debug, Default, Pread)]
+#[derive(Clone, Copy, Debug, Default)]
 struct DebugLinesBlockHeader {
     /// Offset of the file checksum in the file checksums debug subsection.
     file_index: u32,
@@ -398,6 +468,21 @@ struct DebugLinesBlockHeader {
 
     /// Total byte size of this block, including following line and column entries.
     block_size: u32,
+}
+
+impl<'t> TryFromCtx<'t, Endian> for DebugLinesBlockHeader {
+    type Error = scroll::Error;
+    type Size = usize;
+
+    fn try_from_ctx(this: &'t [u8], le: Endian) -> scroll::Result<(Self, Self::Size)> {
+        let mut offset = 0;
+        let data = DebugLinesBlockHeader {
+            file_index: this.gread_with(&mut offset, le)?,
+            num_lines: this.gread_with(&mut offset, le)?,
+            block_size: this.gread_with(&mut offset, le)?,
+        };
+        Ok((data, offset))
+    }
 }
 
 impl DebugLinesBlockHeader {
@@ -509,11 +594,26 @@ impl FileChecksumKind {
 }
 
 /// Raw header of a single file checksum entry.
-#[derive(Clone, Copy, Debug, Pread)]
+#[derive(Clone, Copy, Debug)]
 struct FileChecksumHeader {
     name_offset: u32,
     checksum_size: u8,
     checksum_kind: u8,
+}
+
+impl<'t> TryFromCtx<'t, Endian> for FileChecksumHeader {
+    type Error = scroll::Error;
+    type Size = usize;
+
+    fn try_from_ctx(this: &'t [u8], le: Endian) -> scroll::Result<(Self, Self::Size)> {
+        let mut offset = 0;
+        let data = FileChecksumHeader {
+            name_offset: this.gread_with(&mut offset, le)?,
+            checksum_size: this.gread_with(&mut offset, le)?,
+            checksum_kind: this.gread_with(&mut offset, le)?,
+        };
+        Ok((data, offset))
+    }
 }
 
 /// A file checksum entry.